+++ models/shopacctchange.js
@@ -6,17 +6,17 @@ module.exports = (sequelize, DataTypes) => {
       autoIncrement: true
     },
     CustomedPoints: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
       defaultValue:0
     },
     ChargedMoney: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
             defaultValue:0
     },
     CustomedMoney: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
       defaultValue:0
     },
@@ -25,12 +25,12 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     ShopBounusPoints: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
             defaultValue:0
     },
     RecommendPoints: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.INTEGER,
       allowNull: false,
       defaultValue:0
     },
